Moved endpoint description
